+++ packages/press-ui/src/packages/common/vue3/set.d.ts
@@ -1,0 +1,1 @@
+export function setAdapter<T>(target?: T, key?: any, value?: any): T;