+++ packages/press-ui/src/packages/press-q-r-code/press-q-r-code.vue
@@ -97,6 +97,14 @@ export default {
         });
       },
     },
+    size: {
+      handler() {
+        this.codeImg = '';
+        setTimeout(() => {
+          this.genQRCode();
+        });
+      },
+    },
   },
   mounted() {
     this.genQRCode();